Style fixes to tests
